Handle case where outshine-promotion-headings is unset for org-mode
Suppose we rely on autoloading `navi`, and we open an `org-mode` file. If we try
calling (the autoloaded function) `navi-search-and-switch`, we'll get a cryptic
error about strings. This is because the local variable
`outshine-promotion-headings` has not been set, and `navi` needs this to
construct its regexp. Normally, this is handled by the addition of
`navi-make-org-mode-promotion-headings-list` to the `org-mode-hook`; however,
this doesn't occur until *after* `navi-mode` has been loaded, which
happens *after* we loaded our org-mode file.

This commit fixes this by adding some logic to
`navi-calc-org-mode-headline-regexp` such that if `outshine-promotion-headings`
is nil, it calls `navi-make-org-mode-promotion-headings-list` itself to ensure
that `outshine-promotion-headings` has been set.

In `navi-mode`'s defense, the documentation does suggest including
```
(require 'navi-mode)
```
in the user's config. If this advice is followed, then `navi-mode` should
already have manipulated the `org-mode-hook` prior to opening any `org-mode`
files and we should be all set. However, since `navi-mode` already uses some
autoloading, it seemed natural to make it more robust.